addMethods

+++ src/main/resources/db/migration/V1__init.sql
@@ -1,30 +1,30 @@
 CREATE TABLE IF NOT EXISTS transporters (
-    id INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id bigint PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     name VARCHAR NOT NULL,
     phone VARCHAR(11) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS users (
-    id integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id bigint PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     login VARCHAR NOT NULL UNIQUE,
     password VARCHAR NOT NULL,
     full_name VARCHAR not NULL
 );
 
 CREATE TABLE IF NOT EXISTS trails (
-    id INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id bigint PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     point_departure VARCHAR(50) NOT NULL,
     point_arrival VARCHAR(50) NOT NULL,
     duration INTeger,
-    transporter_id integer REFERENCES transporters(id)
+    transporter_id bigint REFERENCES transporters(id)
 );
 
 CREATE TABLE IF NOT EXISTS tickets (
-    id integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id bigint PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     departure_time DATE NOT NULL,
     place integer NOT NULL,
     price integer NOT NULL,
     paid boolean,
-    user_id integer REFERENCES users(id),
-    trail_id integer REFERENCES trails(id)
+    user_id bigint REFERENCES users(id),
+    trail_id bigint REFERENCES trails(id)
 );
